backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langgraph
openai
python-dotenv
pydantic
pytest
